refactor: organized react hooks
