+++ src/renderer/src/hooks/data/useRunnerLookup.tsx
@@ -1,6 +1,6 @@
 import { useMutation } from "@tanstack/react-query";
 import { useToasts } from "~/features/Toasts/useToasts";
-import { useIpcRenderer } from "./useIpcRenderer";
+import { useIpcRenderer } from "../useIpcRenderer";
 
 export function useRunnerLookup() {
   const ipcRenderer = useIpcRenderer();
